add clinic signup

+++ Backend/src/utils/cloudinary.js
@@ -1,14 +1,12 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
-
-cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-  api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET 
-});
-
 const uploadOnCloudinary = async (localFilePath) => {
+    cloudinary.config({ 
+      cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+      api_key: process.env.CLOUDINARY_API_KEY, 
+      api_secret: process.env.CLOUDINARY_API_SECRET 
+    });
     try {
         if (!localFilePath) return null
         const response = await cloudinary.uploader.upload(localFilePath, {
@@ -17,10 +15,13 @@ const uploadOnCloudinary = async (localFilePath) => {
         return response;
 
     } catch (error) {
+        console.log(error);
         return null;
     }
     finally{
-        fs.unlinkSync(localFilePath);
+        if (fs.existsSync(localFilePath)) {
+          fs.unlinkSync(localFilePath);
+        }
     }
 }
